# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Staffer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gmail/connect": {
      "post": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Connect",
        "operationId": "connect_api_v1_gmail_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GmailConnect"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/gmail/status": {
      "get": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Status",
        "operationId": "status_api_v1_gmail_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/gmail/connection": {
      "delete": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Disconnect",
        "operationId": "disconnect_api_v1_gmail_connection_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/gmail/sync": {
      "post": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Request Sync",
        "operationId": "request_sync_api_v1_gmail_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews": {
      "get": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Reviews",
        "operationId": "reviews_api_v1_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "OPEN",
                "RESOLVED",
                "DISMISSED"
              ],
              "type": "string",
              "default": "OPEN",
              "title": "State"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}": {
      "get": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Review Detail",
        "operationId": "review_detail_api_v1_reviews__review_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/resolve": {
      "post": {
        "tags": [
          "gmail",
          "email review"
        ],
        "summary": "Resolve Review",
        "operationId": "resolve_review_api_v1_reviews__review_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GmailConnect": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "platform": {
            "type": "string",
            "enum": [
              "WINDOWS",
              "ANDROID",
              "windows",
              "android"
            ],
            "title": "Platform"
          },
          "device_label": {
            "type": "string",
            "maxLength": 255,
            "title": "Device Label",
            "default": "Personal device"
          },
          "challenge": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "title": "Challenge"
          },
          "selected_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Label"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "platform",
          "challenge"
        ],
        "title": "GmailConnect"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualApplicationInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Title"
          },
          "company": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Company"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 250000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "notes": {
            "type": "string",
            "maxLength": 50000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "company"
        ],
        "title": "ManualApplicationInput"
      },
...
      "ResolveReview": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "DISMISS",
              "LINK",
              "CREATE_APPLICATION"
            ],
            "title": "Action"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "application_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "application_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Revision"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "APPLIED",
                  "ASSESSMENT",
                  "INTERVIEWING",
                  "OFFER",
                  "REJECTED"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          },
          "application": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualApplicationInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "action",
          "expected_revision",
          "reason"
        ],
        "title": "ResolveReview"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}